# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Torrent streaming server.",
        "title": "Swagger Torrserver API",
        "contact": {},
        "license": {
            "name": "GPL 3.0"
        },
        "version": "{version.Version}"
    },
    "basePath": "/",
    "paths": {
...
        "/torrent/upload": {
            "post": {
                "description": "Only one file support.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Add .torrent file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Torrent file to insert",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Save to DB",
                        "name": "save",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Torrent title",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Torrent category",
                        "name": "category",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Torrent poster",
                        "name": "poster",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Torrent data",
                        "name": "data",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Torrent status",
                        "schema": {
                            "$ref": "#/definitions/state.TorrentStatus"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "state.TorrentFileStat": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "length": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "state.TorrentStat": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5
            ],
            "x-enum-varnames": [
                "TorrentAdded",
                "TorrentGettingInfo",
                "TorrentPreload",
                "TorrentWorking",
                "TorrentClosed",
                "TorrentInDB"
            ]
        },
        "state.TorrentStatus": {
            "type": "object",
            "properties": {
                "active_peers": {
                    "type": "integer"
                },
                "bit_rate": {
                    "type": "string"
                },
                "bytes_read": {
                    "type": "integer"
                },
                "bytes_read_data": {
                    "type": "integer"
                },
                "bytes_read_useful_data": {
                    "type": "integer"
                },
                "bytes_written": {
                    "type": "integer"
                },
                "bytes_written_data": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "chunks_read": {
                    "type": "integer"
                },
                "chunks_read_useful": {
                    "type": "integer"
                },
                "chunks_read_wasted": {
                    "type": "integer"
                },
                "chunks_written": {
                    "type": "integer"
                },
                "connected_seeders": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "download_speed": {
                    "type": "number"
                },
                "duration_seconds": {
                    "type": "number"
                },
                "file_stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/state.TorrentFileStat"
                    }
                },
                "half_open_peers": {
                    "type": "integer"
                },
                "hash": {
                    "type": "string"
                },
                "loaded_size": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pending_peers": {
                    "type": "integer"
                },
                "pieces_dirtied_bad": {
                    "type": "integer"
                },
                "pieces_dirtied_good": {
                    "type": "integer"
                },
                "poster": {
                    "type": "string"
                },
                "preload_size": {
                    "type": "integer"
                },
                "preloaded_bytes": {
                    "type": "integer"
                },
                "stat": {
                    "$ref": "#/definitions/state.TorrentStat"
                },
                "stat_string": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "torrent_size": {
                    "type": "integer"
                },
                "total_peers": {
                    "type": "integer"
                },
                "upload_speed": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}